LPD-27759 Rename existing classes
