--- modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/upgrade/v3_9_0/test/DDMDataProviderInstanceUpgradeProcessTest.java
@@ -1,121 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.dynamic.data.mapping.upgrade.v3_9_0.test;
-
-import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.change.tracking.test.util.BaseCTUpgradeProcessTestCase;
-import com.liferay.dynamic.data.mapping.model.DDMDataProviderInstance;
-import com.liferay.dynamic.data.mapping.model.DDMForm;
-import com.liferay.dynamic.data.mapping.service.DDMDataProviderInstanceLocalService;
-import com.liferay.dynamic.data.mapping.test.util.DDMFormTestUtil;
-import com.liferay.dynamic.data.mapping.test.util.DDMFormValuesTestUtil;
-import com.liferay.portal.kernel.model.Group;
-import com.liferay.portal.kernel.model.change.tracking.CTModel;
-import com.liferay.portal.kernel.service.ServiceContext;
-import com.liferay.portal.kernel.service.change.tracking.CTService;
-import com.liferay.portal.kernel.test.rule.AggregateTestRule;
-import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
-import com.liferay.portal.kernel.test.util.GroupTestUtil;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
-import com.liferay.portal.kernel.upgrade.UpgradeProcess;
-import com.liferay.portal.test.rule.Inject;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
-import com.liferay.portal.upgrade.registry.UpgradeStepRegistrator;
-import com.liferay.portal.upgrade.test.util.UpgradeTestUtil;
-
-import org.junit.Before;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.runner.RunWith;
-
-/**
- * @author Paulo Albuquerque
- */
-@RunWith(Arquillian.class)
-public class DDMDataProviderInstanceUpgradeProcessTest
-	extends BaseCTUpgradeProcessTestCase {
-
-	@ClassRule
-	@Rule
-	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new LiferayIntegrationTestRule(),
-			PermissionCheckerMethodTestRule.INSTANCE);
-
-	@Before
-	public void setUp() throws Exception {
-		_ddmForm = DDMFormTestUtil.createDDMForm(RandomTestUtil.randomString());
-
-		_group = GroupTestUtil.addGroup();
-
-		_serviceContext = ServiceContextTestUtil.getServiceContext(
-			_group.getGroupId());
-	}
-
-	@Override
-	protected CTModel<?> addCTModel() throws Exception {
-		_ddmDataProviderInstance =
-			_ddmDataProviderInstanceLocalService.addDataProviderInstance(
-				TestPropsValues.getUserId(), _group.getGroupId(),
-				RandomTestUtil.randomLocaleStringMap(), null,
-				DDMFormValuesTestUtil.createDDMFormValues(_ddmForm), "rest",
-				_serviceContext);
-
-		return _ddmDataProviderInstance;
-	}
-
-	@Override
-	protected CTService<?> getCTService() {
-		return _ddmDataProviderInstanceLocalService;
-	}
-
-	@Override
-	protected void runUpgrade() throws Exception {
-		UpgradeProcess upgradeProcess = UpgradeTestUtil.getUpgradeStep(
-			_upgradeStepRegistrator, _CLASS_NAME);
-
-		upgradeProcess.upgrade();
-	}
-
-	@Override
-	protected CTModel<?> updateCTModel(CTModel<?> ctModel) throws Exception {
-		_ddmDataProviderInstance = (DDMDataProviderInstance)ctModel;
-
-		return _ddmDataProviderInstanceLocalService.updateDataProviderInstance(
-			_ddmDataProviderInstance.getUserId(),
-			_ddmDataProviderInstance.getDataProviderInstanceId(),
-			_ddmDataProviderInstance.getNameMap(), null,
-			DDMFormValuesTestUtil.createDDMFormValues(_ddmForm),
-			_serviceContext);
-	}
-
-	private static final String _CLASS_NAME =
-		"com.liferay.dynamic.data.mapping.internal.upgrade.v3_9_0." +
-			"DDMDataProviderInstanceUpgradeProcess";
-
-	@Inject(
-		filter = "(&(component.name=com.liferay.dynamic.data.mapping.internal.upgrade.registry.DDMServiceUpgradeStepRegistrator))"
-	)
-	private static UpgradeStepRegistrator _upgradeStepRegistrator;
-
-	@DeleteAfterTestRun
-	private DDMDataProviderInstance _ddmDataProviderInstance;
-
-	@Inject
-	private DDMDataProviderInstanceLocalService
-		_ddmDataProviderInstanceLocalService;
-
-	private DDMForm _ddmForm;
-
-	@DeleteAfterTestRun
-	private Group _group;
-
-	private ServiceContext _serviceContext;
-
-}
