+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/upgrade/v4_0_0/test/DDMStructureVersionCTUpgradeProcessTest.java
@@ -3,23 +3,21 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.dynamic.data.mapping.upgrade.v3_10_2.test;
+package com.liferay.dynamic.data.mapping.upgrade.v4_0_0.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.change.tracking.test.util.BaseCTUpgradeProcessTestCase;
-import com.liferay.dynamic.data.mapping.model.DDMContent;
 import com.liferay.dynamic.data.mapping.model.DDMForm;
 import com.liferay.dynamic.data.mapping.model.DDMFormField;
-import com.liferay.dynamic.data.mapping.model.DDMFormInstance;
-import com.liferay.dynamic.data.mapping.model.DDMFormInstanceRecord;
-import com.liferay.dynamic.data.mapping.model.DDMFormInstanceRecordVersion;
-import com.liferay.dynamic.data.mapping.service.DDMContentLocalService;
-import com.liferay.dynamic.data.mapping.service.DDMFormInstanceRecordVersionLocalService;
-import com.liferay.dynamic.data.mapping.test.util.DDMFormInstanceRecordTestUtil;
-import com.liferay.dynamic.data.mapping.test.util.DDMFormInstanceTestUtil;
+import com.liferay.dynamic.data.mapping.model.DDMStructure;
+import com.liferay.dynamic.data.mapping.model.DDMStructureVersion;
+import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
+import com.liferay.dynamic.data.mapping.service.DDMStructureVersionLocalService;
 import com.liferay.dynamic.data.mapping.test.util.DDMFormTestUtil;
-import com.liferay.dynamic.data.mapping.test.util.DDMFormValuesTestUtil;
+import com.liferay.dynamic.data.mapping.test.util.DDMStructureTestUtil;
+import com.liferay.journal.model.JournalArticle;
 import com.liferay.portal.kernel.json.JSONArray;
+import com.liferay.portal.kernel.json.JSONFactory;
 import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.model.Group;
@@ -28,10 +26,8 @@ import com.liferay.portal.kernel.service.change.tracking.CTService;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.util.GroupTestUtil;
-import com.liferay.portal.kernel.test.util.RandomTestUtil;
-import com.liferay.portal.kernel.test.util.ServiceContextTestUtil;
-import com.liferay.portal.kernel.test.util.TestPropsValues;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
+import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.test.rule.Inject;
 import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
 import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
@@ -47,7 +43,8 @@ import org.junit.runner.RunWith;
  * @author David Truong
  */
 @RunWith(Arquillian.class)
-public class DDMContentUpgradeProcessTest extends BaseCTUpgradeProcessTestCase {
+public class DDMStructureVersionCTUpgradeProcessTest
+	extends BaseCTUpgradeProcessTestCase {
 
 	@ClassRule
 	@Rule
@@ -93,51 +90,18 @@ public class DDMContentUpgradeProcessTest extends BaseCTUpgradeProcessTestCase {
 
 		_ddmForm.addDDMFormField(fieldSetDDMFormField);
 
-		_ddmFormInstance = DDMFormInstanceTestUtil.addDDMFormInstance(
-			_ddmForm, _group, TestPropsValues.getUserId());
-
-		_ddmFormInstanceRecord =
-			DDMFormInstanceRecordTestUtil.addDDMFormInstanceRecord(
-				_ddmFormInstance,
-				DDMFormValuesTestUtil.createDDMFormValues(_ddmForm), _group,
-				TestPropsValues.getUserId());
+		_ddmStructure = DDMStructureTestUtil.addStructure(
+			JournalArticle.class.getName(), _ddmForm);
 	}
 
 	@Override
 	protected CTModel<?> addCTModel() throws Exception {
-		JSONObject jsonObject = JSONUtil.put(
-			"name", RandomTestUtil.randomString());
-
-		JSONArray jsonArray = JSONUtil.putAll(
-			JSONUtil.put(
-				"fieldReference", RandomTestUtil.randomString()
-			).put(
-				"instanceId", RandomTestUtil.randomString()
-			).put(
-				"name", RandomTestUtil.randomString()
-			));
-
-		jsonObject.put("fieldValues", jsonArray);
-
-		DDMContent ddmContent = _ddmContentLocalService.addContent(
-			TestPropsValues.getUserId(), _group.getGroupId(),
-			RandomTestUtil.randomString(), RandomTestUtil.randomString(),
-			jsonObject.toString(), ServiceContextTestUtil.getServiceContext());
-
-		DDMFormInstanceRecordVersion ddmFormInstanceRecordVersion =
-			_ddmFormInstanceRecord.getFormInstanceRecordVersion();
-
-		ddmFormInstanceRecordVersion.setStorageId(ddmContent.getContentId());
-
-		_ddmFormInstanceRecordVersionLocalService.
-			updateDDMFormInstanceRecordVersion(ddmFormInstanceRecordVersion);
-
-		return ddmContent;
+		return _ddmStructure.getStructureVersion();
 	}
 
 	@Override
 	protected CTService<?> getCTService() {
-		return _ddmContentLocalService;
+		return _ddmStructureVersionLocalService;
 	}
 
 	@Override
@@ -150,51 +114,50 @@ public class DDMContentUpgradeProcessTest extends BaseCTUpgradeProcessTestCase {
 
 	@Override
 	protected CTModel<?> updateCTModel(CTModel<?> ctModel) throws Exception {
-		DDMContent ddmContent = (DDMContent)ctModel;
+		DDMStructureVersion ddmStructureVersion = (DDMStructureVersion)ctModel;
 
-		ddmContent.setData(
-			JSONUtil.put(
-				"fieldValues",
-				JSONUtil.putAll(
-					JSONUtil.put(
-						"fieldReference", RandomTestUtil.randomString()
-					).put(
-						"instanceId", RandomTestUtil.randomString()
-					).put(
-						"name", RandomTestUtil.randomString()
-					))
-			).put(
-				"name", RandomTestUtil.randomString()
-			).toString());
+		JSONObject definitionJSONObject = _jsonFactory.createJSONObject(
+			ddmStructureVersion.getDefinition());
 
-		return _ddmContentLocalService.updateDDMContent(ddmContent);
+		JSONArray fieldsJSONArray = definitionJSONObject.getJSONArray("fields");
+
+		JSONObject jsonObject = fieldsJSONArray.getJSONObject(0);
+
+		jsonObject.put("required", true);
+
+		ddmStructureVersion.setDescription(definitionJSONObject.toString());
+
+		return _ddmStructureVersionLocalService.updateDDMStructureVersion(
+			ddmStructureVersion);
 	}
 
 	private static final String _CLASS_NAME =
-		"com.liferay.dynamic.data.mapping.internal.upgrade.v3_10_2." +
-			"DDMContentUpgradeProcess";
+		"com.liferay.dynamic.data.mapping.internal.upgrade.v4_0_0." +
+			"DDMStructureUpgradeProcess";
 
 	@Inject(
 		filter = "(&(component.name=com.liferay.dynamic.data.mapping.internal.upgrade.registry.DDMServiceUpgradeStepRegistrator))"
 	)
 	private static UpgradeStepRegistrator _upgradeStepRegistrator;
 
-	@Inject
-	private DDMContentLocalService _ddmContentLocalService;
-
 	private DDMForm _ddmForm;
 
 	@DeleteAfterTestRun
-	private DDMFormInstance _ddmFormInstance;
-
-	@DeleteAfterTestRun
-	private DDMFormInstanceRecord _ddmFormInstanceRecord;
+	private DDMStructure _ddmStructure;
 
 	@Inject
-	private DDMFormInstanceRecordVersionLocalService
-		_ddmFormInstanceRecordVersionLocalService;
+	private DDMStructureLocalService _ddmStructureLocalService;
+
+	@Inject
+	private DDMStructureVersionLocalService _ddmStructureVersionLocalService;
 
 	@DeleteAfterTestRun
 	private Group _group;
+
+	@Inject
+	private JSONFactory _jsonFactory;
+
+	@Inject
+	private Portal _portal;
 
 }
