+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/upgrade/v4_3_2/test/DDMTemplateCTUpgradeProcessTest.java
@@ -3,29 +3,24 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.dynamic.data.mapping.upgrade.v4_0_0.test;
+package com.liferay.dynamic.data.mapping.upgrade.v4_3_2.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
 import com.liferay.change.tracking.test.util.BaseCTUpgradeProcessTestCase;
-import com.liferay.dynamic.data.mapping.model.DDMForm;
-import com.liferay.dynamic.data.mapping.model.DDMFormField;
 import com.liferay.dynamic.data.mapping.model.DDMStructure;
-import com.liferay.dynamic.data.mapping.model.DDMStructureVersion;
-import com.liferay.dynamic.data.mapping.service.DDMStructureLocalService;
-import com.liferay.dynamic.data.mapping.service.DDMStructureVersionLocalService;
-import com.liferay.dynamic.data.mapping.test.util.DDMFormTestUtil;
+import com.liferay.dynamic.data.mapping.model.DDMTemplate;
+import com.liferay.dynamic.data.mapping.service.DDMTemplateLocalService;
 import com.liferay.dynamic.data.mapping.test.util.DDMStructureTestUtil;
+import com.liferay.dynamic.data.mapping.test.util.DDMTemplateTestUtil;
 import com.liferay.journal.model.JournalArticle;
-import com.liferay.portal.kernel.json.JSONArray;
 import com.liferay.portal.kernel.json.JSONFactory;
-import com.liferay.portal.kernel.json.JSONObject;
-import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.change.tracking.CTModel;
 import com.liferay.portal.kernel.service.change.tracking.CTService;
 import com.liferay.portal.kernel.test.rule.AggregateTestRule;
 import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.kernel.test.util.GroupTestUtil;
+import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
 import com.liferay.portal.kernel.util.Portal;
 import com.liferay.portal.test.rule.Inject;
@@ -43,7 +38,7 @@ import org.junit.runner.RunWith;
  * @author David Truong
  */
 @RunWith(Arquillian.class)
-public class DDMStructureUpgradeProcessTest
+public class DDMTemplateCTUpgradeProcessTest
 	extends BaseCTUpgradeProcessTestCase {
 
 	@ClassRule
@@ -57,51 +52,20 @@ public class DDMStructureUpgradeProcessTest
 	public void setUp() throws Exception {
 		_group = GroupTestUtil.addGroup();
 
-		_ddmForm = DDMFormTestUtil.createDDMForm();
-
-		DDMFormField fieldSetDDMFormField = DDMFormTestUtil.createDDMFormField(
-			"fieldset", "fieldset", "fieldset", "", false, false, false);
-
-		fieldSetDDMFormField.addNestedDDMFormField(
-			DDMFormTestUtil.createTextDDMFormField(
-				"field1", false, false, false));
-		fieldSetDDMFormField.addNestedDDMFormField(
-			DDMFormTestUtil.createTextDDMFormField(
-				"field2", false, false, false));
-		fieldSetDDMFormField.setProperty(
-			"rows",
-			JSONUtil.putAll(
-				JSONUtil.put(
-					"columns",
-					JSONUtil.put(
-						JSONUtil.put(
-							"fields", JSONUtil.put("field1")
-						).put(
-							"size", 12
-						))),
-				JSONUtil.put(
-					"columns",
-					JSONUtil.put(
-						JSONUtil.put(
-							"fields", JSONUtil.put("field2")
-						).put(
-							"size", 12
-						)))));
-
-		_ddmForm.addDDMFormField(fieldSetDDMFormField);
-
 		_ddmStructure = DDMStructureTestUtil.addStructure(
-			JournalArticle.class.getName(), _ddmForm);
+			_group.getGroupId(), JournalArticle.class.getName());
 	}
 
 	@Override
 	protected CTModel<?> addCTModel() throws Exception {
-		return _ddmStructure.getStructureVersion();
+		return DDMTemplateTestUtil.addTemplate(
+			_ddmStructure.getStructureId(),
+			_portal.getClassNameId(JournalArticle.class.getName()));
 	}
 
 	@Override
 	protected CTService<?> getCTService() {
-		return _ddmStructureVersionLocalService;
+		return _ddmTemplateLocalService;
 	}
 
 	@Override
@@ -114,42 +78,27 @@ public class DDMStructureUpgradeProcessTest
 
 	@Override
 	protected CTModel<?> updateCTModel(CTModel<?> ctModel) throws Exception {
-		DDMStructureVersion ddmStructureVersion = (DDMStructureVersion)ctModel;
+		DDMTemplate ddmTemplate = (DDMTemplate)ctModel;
 
-		JSONObject definitionJSONObject = _jsonFactory.createJSONObject(
-			ddmStructureVersion.getDefinition());
+		ddmTemplate.setScript(RandomTestUtil.randomString());
 
-		JSONArray fieldsJSONArray = definitionJSONObject.getJSONArray("fields");
-
-		JSONObject jsonObject = fieldsJSONArray.getJSONObject(0);
-
-		jsonObject.put("required", true);
-
-		ddmStructureVersion.setDescription(definitionJSONObject.toString());
-
-		return _ddmStructureVersionLocalService.updateDDMStructureVersion(
-			ddmStructureVersion);
+		return _ddmTemplateLocalService.updateDDMTemplate(ddmTemplate);
 	}
 
 	private static final String _CLASS_NAME =
-		"com.liferay.dynamic.data.mapping.internal.upgrade.v4_0_0." +
-			"DDMStructureUpgradeProcess";
+		"com.liferay.dynamic.data.mapping.internal.upgrade.v4_3_2." +
+			"DDMTemplateUpgradeProcess";
 
 	@Inject(
 		filter = "(&(component.name=com.liferay.dynamic.data.mapping.internal.upgrade.registry.DDMServiceUpgradeStepRegistrator))"
 	)
 	private static UpgradeStepRegistrator _upgradeStepRegistrator;
 
-	private DDMForm _ddmForm;
-
 	@DeleteAfterTestRun
 	private DDMStructure _ddmStructure;
 
 	@Inject
-	private DDMStructureLocalService _ddmStructureLocalService;
-
-	@Inject
-	private DDMStructureVersionLocalService _ddmStructureVersionLocalService;
+	private DDMTemplateLocalService _ddmTemplateLocalService;
 
 	@DeleteAfterTestRun
 	private Group _group;
