+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/upgrade/v4_3_3/test/DDMStructureLayoutCTUpgradeProcessTest.java
@@ -15,7 +15,7 @@ import org.junit.runner.RunWith;
  * @author Paulo Albuquerque
  */
 @RunWith(Arquillian.class)
-public class DDMStructureLayoutUpgradeProcessTest
+public class DDMStructureLayoutCTUpgradeProcessTest
 	extends BaseDDMStructureLayoutCTUpgradeProcessTestCase {
 
 	@Override
