+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/upgrade/v5_1_0/test/DDMFieldCTUpgradeProcessTest.java
@@ -56,7 +56,7 @@ import org.junit.runner.RunWith;
  * @author Paulo Albuquerque
  */
 @RunWith(Arquillian.class)
-public class DDMFieldUpgradeProcessTest extends BaseCTUpgradeProcessTestCase {
+public class DDMFieldCTUpgradeProcessTest extends BaseCTUpgradeProcessTestCase {
 
 	@ClassRule
 	@Rule
