+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/upgrade/v5_1_0/test/DDMStorageLinkDDMFieldCTUpgradeProcessTest.java
@@ -22,8 +22,8 @@ import org.junit.runner.RunWith;
  * @author Paulo Albuquerque
  */
 @RunWith(Arquillian.class)
-public class DDMStorageLinkDDMFieldUpgradeProcessTest
-	extends DDMFieldUpgradeProcessTest {
+public class DDMStorageLinkDDMFieldCTUpgradeProcessTest
+	extends DDMFieldCTUpgradeProcessTest {
 
 	@ClassRule
 	@Rule
