+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/upgrade/v5_1_5/test/DDMStructureLayoutCTUpgradeProcessTest.java
@@ -14,7 +14,7 @@ import org.junit.runner.RunWith;
  * @author Paulo Albuquerque
  */
 @RunWith(Arquillian.class)
-public class DDMStructureLayoutUpgradeProcessTest
+public class DDMStructureLayoutCTUpgradeProcessTest
 	extends BaseDDMStructureLayoutCTUpgradeProcessTestCase {
 
 	@Override
