+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-test/src/testIntegration/java/com/liferay/dynamic/data/mapping/upgrade/v5_3_2/test/DDMFormInstanceVersionDDMFormInstanceCTUpgradeProcessTest.java
@@ -3,39 +3,26 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.dynamic.data.mapping.upgrade.v4_3_1.test;
+package com.liferay.dynamic.data.mapping.upgrade.v5_3_2.test;
 
 import com.liferay.arquillian.extension.junit.bridge.junit.Arquillian;
-import com.liferay.dynamic.data.mapping.io.DDMFormValuesSerializer;
 import com.liferay.dynamic.data.mapping.model.DDMFormInstance;
 import com.liferay.dynamic.data.mapping.model.DDMFormInstanceVersion;
 import com.liferay.dynamic.data.mapping.service.DDMFormInstanceVersionLocalService;
-import com.liferay.dynamic.data.mapping.test.util.DDMFormInstanceTestUtil;
-import com.liferay.dynamic.data.mapping.util.DDMFormValuesSerializeUtil;
+import com.liferay.portal.kernel.model.Group;
 import com.liferay.portal.kernel.model.change.tracking.CTModel;
 import com.liferay.portal.kernel.service.change.tracking.CTService;
-import com.liferay.portal.kernel.test.rule.AggregateTestRule;
+import com.liferay.portal.kernel.test.rule.DeleteAfterTestRun;
 import com.liferay.portal.test.rule.Inject;
-import com.liferay.portal.test.rule.LiferayIntegrationTestRule;
-import com.liferay.portal.test.rule.PermissionCheckerMethodTestRule;
 
-import org.junit.ClassRule;
-import org.junit.Rule;
 import org.junit.runner.RunWith;
 
 /**
  * @author Paulo Albuquerque
  */
 @RunWith(Arquillian.class)
-public class DDMFormInstanceVersionDDMFormInstanceUpgradeProcessTest
-	extends DDMFormInstanceUpgradeProcessTest {
-
-	@ClassRule
-	@Rule
-	public static final AggregateTestRule aggregateTestRule =
-		new AggregateTestRule(
-			new LiferayIntegrationTestRule(),
-			PermissionCheckerMethodTestRule.INSTANCE);
+public class DDMFormInstanceVersionDDMFormInstanceCTUpgradeProcessTest
+	extends DDMFormInstanceCTUpgradeProcessTest {
 
 	@Override
 	protected CTModel<?> addCTModel() throws Exception {
@@ -55,20 +42,23 @@ public class DDMFormInstanceVersionDDMFormInstanceUpgradeProcessTest
 		DDMFormInstanceVersion ddmFormInstanceVersion =
 			(DDMFormInstanceVersion)ctModel;
 
-		ddmFormInstanceVersion.setSettings(
-			DDMFormValuesSerializeUtil.serialize(
-				DDMFormInstanceTestUtil.createSettingsDDMFormValues(),
-				_jsonDDMFormValuesSerializer));
+		DDMFormInstance ddmFormInstance = updateDDMFormInstance(
+			ddmFormInstanceVersion.getFormInstance());
+
+		ddmFormInstanceVersion.setSettings(ddmFormInstance.getSettings());
 
 		return _ddmFormInstanceVersionLocalService.updateDDMFormInstanceVersion(
 			ddmFormInstanceVersion);
 	}
 
+	@DeleteAfterTestRun
+	private DDMFormInstance _ddmFormInstance;
+
 	@Inject
 	private DDMFormInstanceVersionLocalService
 		_ddmFormInstanceVersionLocalService;
 
-	@Inject(filter = "ddm.form.values.serializer.type=json")
-	private DDMFormValuesSerializer _jsonDDMFormValuesSerializer;
+	@DeleteAfterTestRun
+	private Group _group;
 
 }
